# docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  entity.User:
    properties:
      address:
        type: string
      age:
        type: integer
      code:
        type: string
      name:
        type: string
    type: object
  vo.LoginJSON:
    properties:
      password:
        type: string
      user:
        type: string
    required:
    - password
    - user
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /auth/secret:
    get:
      consumes:
      - application/json
      description: 在header 头增加 ["Authorization":"Basic Zm9vOmJhcg=="],["Authorization":"Basic
        YWRtaW46cGFzc3dvcmQ="]
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: Basic auth 使用方法
  /v1/login:
    post:
      consumes:
      - application/json
      description: user login  with username and pwd
      parameters:
      - description: user body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/vo.LoginJSON'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: user login with username and pwd
  /v1/users:
    get:
      consumes:
      - application/json
      description: list all users
      produces:
      - application/json
      responses:
        "200":
          description: '[]entity.User'
          schema:
            $ref: '#/definitions/app.Response'
            type: object
      summary: list all users
  /v1/users/{name}:
    get:
      consumes:
      - application/json
      description: find user  by name
      parameters:
      - description: user name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
            type: object
      summary: find user by name
  /v2/submit:
    get:
      consumes:
      - application/json
      description: 测试数据：router 的另一中写法
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
            type: object
      summary: router 的另一种写法
swagger: "2.0"